# backend/app/Interfaces/Http/OpenApi/Schemas/Requests/UserStoreRequestSchema.php
<?php

namespace App\Interfaces\Http\OpenApi\Schemas\Requests;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'UserStoreRequest',
    type: 'object',
    required: ['name', 'email', 'password'],
    properties: [
        new OA\Property(
            property: 'name',
            type: 'string',
            maxLength: 255,
            example: 'Igor'
        ),
        new OA\Property(
            property: 'email',
            type: 'string',
            format: 'email',
            maxLength: 255,
            example: 'igor@email.com'
        ),
        new OA\Property(
            property: 'password',
            type: 'string',
            format: 'password',
            minLength: 8,
            example: '12345678'
        ),
    ]
)]
class UserStoreRequestSchema
{
}