# backend/app/Interfaces/Http/OpenApi/Schemas/Responses/UserResponseSchema.php
<?php

namespace App\Interfaces\Http\OpenApi\Schemas\Responses;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'UserResponse',
    type: 'object',
    properties: [
        new OA\Property(
            property: 'id',
            type: 'integer',
            example: 1
        ),
        new OA\Property(
            property: 'name',
            type: 'string',
            example: 'Admin'
        ),
        new OA\Property(
            property: 'email',
            type: 'string',
            format: 'email',
            example: 'admin@example.com'
        ),
        new OA\Property(
            property: 'email_verified_at',
            type: 'string',
            format: 'date-time',
            nullable: true,
            example: null
        ),
        new OA\Property(
            property: 'created_at',
            type: 'string',
            format: 'date-time',
            example: '2026-08-14T02:35:02.000000Z'
        ),
        new OA\Property(
            property: 'updated_at',
            type: 'string',
            format: 'date-time',
            example: '2026-08-14T02:35:02.000000Z'
        ),
        new OA\Property(
            property: 'profiles',
            type: 'array',
            items: new OA\Items(
                ref: '#/components/schemas/ProfileResponse'
            )
        ),
    ]
)]
class UserResponseSchema
{
}